requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==4.1.9
djangorestframework==3.14.0
djoser==2.2.0
mysqlclient==2.1.1
mysql-python
